[Refactor] Redis 폴더를 global 디렉토리로 이동
