+++ src/main/java/org/spring/dojooo/auth/jwt/security/SecurityConfig.java
@@ -2,7 +2,7 @@ package org.spring.dojooo.auth.jwt.security;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.spring.dojooo.auth.Redis.RedisUtil;
+import org.spring.dojooo.global.Redis.RedisUtil;
 import org.spring.dojooo.auth.jwt.config.JWTUtil;
 import org.spring.dojooo.auth.jwt.filter.JWTFilter;
 import org.spring.dojooo.auth.jwt.filter.LoginFilter;
